replacing nemesis
